Домашняя работа 5.12. Разбил код на модули
